+++ src/components/API/GraphQL/queries.generated.tsx
@@ -7,7 +7,7 @@ const defaultOptions =  {}
 export type GetConfigurationQueryVariables = Types.Exact<{ [key: string]: never; }>;
 
 
-export type GetConfigurationQuery = { __typename?: 'Query', configuration: { __typename?: 'Configuration', connectorId: string, hasGalaxy?: Types.Maybe<boolean>, hasGrouping?: Types.Maybe<boolean>, hasFilters?: Types.Maybe<boolean>, enableSSO?: Types.Maybe<boolean>, skipAuth?: Types.Maybe<boolean>, skipTos?: Types.Maybe<boolean>, version: string } };
+export type GetConfigurationQuery = { __typename?: 'Query', configuration: { __typename?: 'Configuration', connectorId: string, hasGalaxy?: Types.Maybe<boolean>, hasFilters?: Types.Maybe<boolean>, hasGrouping?: Types.Maybe<boolean>, enableSSO?: Types.Maybe<boolean>, skipAuth?: Types.Maybe<boolean>, skipTos?: Types.Maybe<boolean>, version: string } };
 
 export type ListAlgorithmsQueryVariables = Types.Exact<{ [key: string]: never; }>;
 
@@ -126,6 +126,7 @@ export const GetConfigurationDocument = gql`
     hasGalaxy
     hasGrouping
     hasFilters
+    hasGrouping
     enableSSO
     skipAuth
     skipTos